Refactor pt. 2
